object-detection/requirements.txt
tensorflow-cpu >= 2.0.0, <= 3.0.0
tensorflow-hub
Flask >= 1.1.2, < 2.0.0
gunicorn >= 20.0.4, < 21.0.0
opencv-python
